fixed numberous bugs with zombie culling; removed redundant '--crtstatic'

+++ src/exec/prep.rs
@@ -1,10 +1,10 @@
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 
 pub fn ensure_out_dirs(sdir: &Path, odir: &Path) {
     let _ = std::fs::create_dir_all(odir);
     let _ = std::fs::create_dir(odir.join("pch"));
-    ensure_out_dirs_rec(&PathBuf::from(sdir), sdir, &odir.join("obj"));
+    ensure_out_dirs_rec(sdir, sdir, &odir.join("obj"));
 }
 
 fn ensure_out_dirs_rec(root: &Path, sdir: &Path, odir: &Path) {
@@ -26,12 +26,13 @@ fn cull_zombies_rec(root: &Path, sdir: &Path, odir: &Path, ext: &str) {
     for e in std::fs::read_dir(root).ok().unwrap() {
         let e = e.ok().unwrap();
         if e.path().is_file() && e.path().extension().unwrap() == "obj" {
-            let src = sdir.join(e.path().file_stem().unwrap()).with_extension(ext);
+            if e.path().to_string_lossy().ends_with(".h.obj") { continue }
+            let src = sdir.join(e.path().strip_prefix(odir).unwrap()).with_extension(ext);
             if !src.exists() {
                 let _ = std::fs::remove_file(e.path());
             }
         } else if e.path().is_dir() {
-            ensure_out_dirs_rec(&e.path(), sdir, odir);
+            cull_zombies_rec(&e.path(), sdir, odir, ext);
         }
     }
 }
